Added paint tools (#1)
* Initial draft of paint tools

* Added additional paint tools; fixed function argument binding

* Added oval, shape and text tool; various code cleanups

* Moving classes

* Added escape modified to polygon tool

* Added icons to tools palette

* Added shapes palette; support for changing text tool font face

* Added provider architecture; refactor and cleanup

* Added checkmarks to font/style menus

* Improved providers

* Additional refactoring

* Palette improvements

* Fixed paint tool issue when changing cards

* Added lines palette

* Added fill tool

* Added patterns

* Added spray paint

* Added curve tool

* Package refactoring

* Fixed warnings originating from ToolsPalette form

* Added default cursors

* Implemented lasso tool

* Tool refactoring

* Improvements to CurveTool

* Improvements to text tool

* Modified documentation for paint tools

+++ src/main/java/hypertalk/ast/functions/ExpAverageFun.java
@@ -37,7 +37,7 @@ public class ExpAverageFun extends ArgListFunction {
                 throw new HtSemanticException("Can't take the average of a non-numerical list");
             }
                 
-            sum += item.floatValue();
+            sum += item.doubleValue();
         }
         
         return new Value(sum/list.size());
