+++ src/main/java/hypertalk/ast/functions/ExpMinFun.java
@@ -4,8 +4,6 @@ import hypertalk.ast.common.Value;
 import hypertalk.ast.expressions.Expression;
 import hypertalk.exception.HtSemanticException;
 
-import java.util.List;
-
 public class ExpMinFun extends ArgListFunction {
 
     public ExpMinFun (Expression expression) {
@@ -26,7 +24,7 @@ public class ExpMinFun extends ArgListFunction {
                 throw new HtSemanticException("All arguments to min() must be numbers.");
             }
 
-            if (thisValue.floatValue() < min.floatValue()) {
+            if (thisValue.doubleValue() < min.doubleValue()) {
                 min = thisValue;
             }
         }
